refactor(discovery): decompose the device registry god-file by role (WS-D)
Split internal/discovery/registry.go (616 LOC) into role-based files in the
same package — pure move: registry type, lifecycle, add/update, and query
accessors stay in registry.go; the granular field-merge helpers move out.

All files under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
